requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
pillow==11.3.0
    # via
    #   -r requirements.in
    #   pystray
psutil==7.0.0
    # via -r requirements.in
pystray==0.19.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
six==1.17.0
    # via
    #   pystray
    #   python-dateutil
watchdog==6.0.0
    # via -r requirements.in
